Add score where the players get 1 point for hitting the other player's wall
Basically made it so that colliding the ball into the wall behind the opposite player grants
a point if the first player was the last player to have hit the ball. Example: Paddle 1 would
receive 1 point for hitting Wall2.

In addition added a live debug textmesh that shows who hit the ball last.