Remade the reward functions, collision handling and cleanup
Now there are two Paddle agents that hopefully will play against each other
after training. The code handling the rewards and punishments in "Ball.cs" has been
rewritten and the debug text has been added back to see each agent's score as they play.
Also rewritten the game to reset whenever a point is scored, as that seems to be the
classic way of playing pong, as opposed to just letting the game continue from there.
This also removes the issue of having the ball suddenly stuck going perfectly vertically or
horizontally as it can only hit one of the sides before the game resets again.