Clamped paddle movement and added a few builds+trained brains
Made it so that the paddles can't leave the play area, as it is
most likely causing the training to break as well. After that I have started
training the brain, however it is unsuccessfull as of now. Everything seems to
be working properly, there is just something wrong with how rewards are assigned it
seems like, as the mean reward keeps beeing 0. Pushing the code up to continue
work at school.