Added scene for Imitation learning
Adding a scene for imitation learning was not as straight forward as it could have been. Since you often want to run many simulations in a paralell in general or at least have a student vs teacher environment, it is important that positions are not hard coded. Resetting the ball and the paddle has been changed to remember their starting positions, so that I can create many games in paralell. This will be beneficial for running many games at once for reinforcement learning (when I get that to work) and imitation learning that I am going to attempt next.