Added two different brains
One brain was trained with trying to get the paddles to hit
the ball more than once, but that failed as many of the games
managed to get the ball to stop moving. The second brain tries to
improve on that by having a 30 second reset, that constantly resets
the game in order to never get stuck forever. However, that brain is
faulty too.