Tried finding a solution for deleting and respawning the "bricks" that are hit on ResetGame()
First tried reloading the scene, but this destroys the communication object with MLAgents.
Then tried to Destroy the prefab and reset the prefab, but that didnt't work.
Current attempt is destroying the prefab and loading the prefab from file,
it works in practise, but after the initial Destroy, the other prefabs
are still in existance, meaning it would eat up RAM quite quickly.
Ended up making a loop that deletes all gameobjects with "Obstacle" in its name, then
loading the 4 different prefabs from file.

This commit includes a 5M iterations model from the last implementation, where
the paddle is good at hitting the ball, but doesn't seem to have developed any
strategical skills beyond that.