Fixed bug where bricks did not respawn on game reset with 0 lives
The bricks would not respawn if the agent managed to clear all the bricks, because
the game reset did not take into account that all of the bricks would be removed.

Now the reset will reset the bricks if all the bricks were removed. A softer reset
also still exists if the agent has more lives left.